requirements.txt
numpy>=1.24
scipy>=1.11
matplotlib>=3.7
pandas>=2.0
pyyaml>=6.0
h5py>=3.9
